Adding placeholder files to 'docs' folder
The files that have been added correspond to the files outlined in the
configuration file:

    mkdocs.yml

All files and the layout is expected to change significantly. These are
mostly placeholder files with dummy contents.

Some of the static documentation files have been created using the package: mkdocs
(https://mkdocs.readthedocs.io/en/stable/)

When one compiles the documentation with:

    mkdocs build

A folder called 'site' is created that contains the static files to be
user for deployment. This is not required to be tracked as one would
rebuild using the files found in 'docs/'.

NOTE. If files are removed from 'docs/' in between builds, they make
still hang around, a better command to run would be:

    mkdocs build --clean

 Changes to be committed:
	new file:   docs/about.md
	new file:   docs/img/overview.png
	new file:   docs/index.md
	new file:   docs/user-guide/getting-started-II.md
	new file:   docs/user-guide/getting-started.md